Completed Task

+++ 0x01-python-if_else_loops_functions/6-print_comb3.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python3
 for i in range(100):
-    if i % 10 > i / 10:
-        print("{}{}".format(i / 10, i % 10), end=", ")
+    if i % 10 > int(i / 10):
+        print("{}{}".format(int(i / 10), i % 10), end=", ")
 print("")
